Created global function to validate paid command usage

+++ commands/transfer.js
@@ -1,11 +1,10 @@
 module.exports = {
     name: 'transfer',
     description: 'allows transfer of gbp from one person to another',
-    execute(message,args,total_money, master){
+    execute(message, args, master){
         try{
             const unlock = require('./Functions/Achievement_Functions')
-            var command
-            var recipient = args[1].toLowerCase() || 'none';
+            const general = require('./Functions/GeneralFunctions')
             var amount = args[2] || 'none';
             if(String(amount).toLowerCase() == 'all'){
                 amount = parseFloat(master[message.author.id].gbp)
@@ -13,60 +12,41 @@ module.exports = {
                 amount = parseFloat(amount)
             }
             
-            for(i in master){
-                if(master[i].name.toLowerCase() == recipient){
-                    var recipient_id = i
-                }
+            //replace this with checking for length of args array
+            if(args.length < 2){
+                message.channel.send("You must specify a person and an amount. ex !transfer [person] [amount]")
+                return;
             }
+
+            var recipient = args[1].toLowerCase();
+            var targetID = general.NameToUserID(args[1].toLowerCase(), master)
 
             if(message.channel.id == 711634711281401867 || message.channel.id == 702205197740540004){
-                command = 'delete'
-            }else if(typeof(recipient_id) !== 'undefined'){
-                command = 'name'
-            }else if(recipient == 'split'){
-                    command = 'split'
-            }else if(typeof(recipient_id) == 'undefined'){
-                command = 'bad_name'
+                message.channel.bulkDelete(1)
+
+                //Your a Dumbass Achievement
+                unlock.unlock(message.author.id, 7, message, master)
+                return
             }
+
+            if(recipient == 'split'){
+                Split(message, master, args, amount)
+                return;
+            }
+
+            if(recipient == "help"){
+                message.channel.send("!transfer [person] [amount] for basic")
+                message.channel.send("!transfer split [amount] [person 1] [person 2] ... to split an amount amongst multiple people")
+                return
+            }
+
+            if(!general.CommandUsageValidator(message, master, amount, 0, master[message.author.id].gbp, targetID)){
+                return
+            }
+
+            general.CommandPurchase(message, master, amount, targetID)
+            message.channel.send(`${master[message.author.id].name} has transferred ${master[targetID].name} ${amount} gbp`)
             
-
-            switch(command){
-                case 'name':
-                    Standard(message, master, args, recipient_id, amount)
-                break;
-                case 'split':
-                    Split(message, master, args, amount)
-                break;
-                case 'bad_name':
-                    message.channel.send(`The recipient doesn't exst`)
-                break;
-                case 'delete':
-                    message.channel.bulkDelete(1)
-
-                    //Your a Dumbass Achievement
-                    unlock.unlock(message.author.id, 7, message, master)
-                break;
-                default: 
-                    message.channel.send('Use "!transfer help" for a list of commands')
-            }
-        }catch(err){
-            console.log(err)
-            message.channel.send('Error occurred in transfer.js')
-        }
-    }
-}
-
-function Standard(message, master, args, recipient_id, amount){
-    try{
-        const unlock = require('./Functions/Achievement_Functions')
-        if(isNaN(amount) == true || amount <= 0){
-            message.channel.send('You must choose a valid amount greater than 0')
-        }else if(amount > master[message.author.id].gbp){
-            message.channel.send(`You can't transfer more gbp than you have`)
-        }else{
-            master[message.author.id].gbp -= amount
-            master[recipient_id].gbp += amount
-            message.channel.send(`${master[message.author.id].name} has transferred ${master[recipient_id].name} ${amount} gbp`)
             if(message.channel.type === 'dm'){
                 var users = message.mentions._client.users.array()
                 for(var k in users){
@@ -81,10 +61,10 @@ function Standard(message, master, args, recipient_id, amount){
                 unlock.unlock(recipient_id, 16, message, master)
             }
 
+        }catch(err){
+            console.log(err)
+            message.channel.send('Error occurred in transfer.js')
         }
-    }catch(err){
-        console.log(err)
-        message.channel.send('Error occurred in transfer.js standard')
     }
 }
